requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-allauth==0.63.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-multiselectfield==0.1.12
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
